+++ src/Application/Domain/IAnyPayload.ts
@@ -1,0 +1,7 @@
+import type IPayload from './IPayload';
+
+type IAnyPayload = {
+	event: string, 
+	payload: IPayload
+}
+export default IAnyPayload;